// src/api/openapi/models/evaluation-detail-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zeroone API v1
 * v1
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface EvaluationDetailResponseDto {
    'evaluationId'?: number;
    'evaluationGrade'?: EvaluationDetailResponseDtoEvaluationGradeEnum;
    'evaluationComment'?: string;
}

export const EvaluationDetailResponseDtoEvaluationGradeEnum = {
    APlus: 'A_PLUS',
    A: 'A',
    BPlus: 'B_PLUS',
    B: 'B',
    CPlus: 'C_PLUS',
    C: 'C',
    DPlus: 'D_PLUS',
    D: 'D',
    F: 'F'
} as const;

export type EvaluationDetailResponseDtoEvaluationGradeEnum = typeof EvaluationDetailResponseDtoEvaluationGradeEnum[keyof typeof EvaluationDetailResponseDtoEvaluationGradeEnum];